pyproject.toml
[project]
name = "word-ocr"
version = "0.1.0"
description = "Extract images from Word documents and OCR to Markdown"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "click>=8.1.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
word-ocr = "word_ocr.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0", "pytest-mock>=3.12.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
